// src/transaction-history/transaction-history.type.ts
export enum TransactionHistoryType {
  PAYMENT = 'payment',
  DISBURSEMENT_NEW_CONTRACT = 'disbursement_new_contract',
  DEDUCTION_MONEY = 'deduction_money',
  PAYMENT_DOWN_ROOT_MONEY = 'payment_down_root_money',
  LOAN_MORE_MONEY = 'loan_more_money',
  SETTLEMENT_CONTRACT = 'settlement_contract',
}